Moved witness internals out a layer
This allows it to be accessible from the omniwitness implementations
