+++ witness/golang/cmd/witness/impl/witness.go
@@ -26,8 +26,8 @@ import (
 	logfmt "github.com/google/trillian-examples/formats/log"
 	i_note "github.com/google/trillian-examples/internal/note"
 	ih "github.com/google/trillian-examples/witness/golang/cmd/witness/internal/http"
-	wsql "github.com/google/trillian-examples/witness/golang/cmd/witness/internal/persistence/sql"
-	"github.com/google/trillian-examples/witness/golang/cmd/witness/internal/witness"
+	wsql "github.com/google/trillian-examples/witness/golang/internal/persistence/sql"
+	"github.com/google/trillian-examples/witness/golang/internal/witness"
 	"github.com/gorilla/mux"
 	_ "github.com/mattn/go-sqlite3" // Load drivers for sqlite3
 	"github.com/transparency-dev/merkle/rfc6962"
