+++ witness/golang/internal/persistence/inmemory/inmemory.go
@@ -16,7 +16,7 @@
 package inmemory
 
 import (
-	"github.com/google/trillian-examples/witness/golang/cmd/witness/internal/persistence"
+	"github.com/google/trillian-examples/witness/golang/internal/persistence"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
