+++ witness/golang/internal/persistence/persistence.go
