+++ witness/golang/internal/persistence/sql/sql.go
@@ -19,7 +19,7 @@ import (
 	"database/sql"
 	"fmt"
 
-	"github.com/google/trillian-examples/witness/golang/cmd/witness/internal/persistence"
+	"github.com/google/trillian-examples/witness/golang/internal/persistence"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
